+++ src/main/java/com/example/my_app/DTO/ship/ShipStatusHistoryDTO.java
@@ -1,16 +1,11 @@
 package com.example.my_app.DTO.ship;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 import com.example.my_app.Configuration.AutoDeserializer;
 import com.example.my_app.DTO.Order.Order_WayBillDTO;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -33,9 +28,9 @@ public class ShipStatusHistoryDTO {
     String shippingPartnerCode;
     String trackingNumber;
     LocalDateTime time;
-    BigDecimal invoiceValue;
-    BigDecimal outstandingCod;
-    BigDecimal shippingFee;
+    int invoiceValue;
+    int outstandingCod;
+    int shippingFee;
     String status;
     ShipmentInfoDTO ShipmentInfo_id;
     Order_WayBillDTO shipStatusHistory_WayBill;
